Create setup-when-root.sh
This is just a quick mockup of approximately how I would do in Ubuntu, don't know if it´s compatible with Vagrant and if it´s even possible. It's just written from my head, and not tested yet.

@jnweiger Please have a look at the code and ask me if you don't get my idea.

What we also need to create is the secure-permisisons.sh script and the .conf file for the self-signed SSL in order for this to even run.

My idea is to create this script during build and then put it in /var/scripts/ inside the VM and then add the last line(s) to /root/profile/ so that shen the user becomes root, this appears. 

The scripts folder is already created in the welcome.sh message and this is supposed to come later. My suggestion is to tell the user to login as the vagrant user the first time and then type sudo-i to become root, and then this script starts.

So in total:

1. user logs in as user vagrant --> presented with welcome.sh
2. user becomes root --> presented with this (or something like it)
3. Setup is done

Files that needs to be created during the build:
- welcome.sh /for instructions
- setup_when_root.sh /for final setup 
- owncloud-self-signed-SSL.conf /SSL
- secure-permissions.sh /secure permissions